git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@26793 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/ipc/shared/networku.c
@@ -18,11 +18,11 @@
 					((x & 0xff00) >> 8))	
 
 unsigned short ntohs(unsigned short x) {
-	return NTOHS(x);
+	return (unsigned short) NTOHS(x);
 }
 
 unsigned short htons(short x) {
-	return NTOHS(x);
+	return (unsigned short) NTOHS(x);
 }
 
 long ntohl(long x) {
